Simpler sparse approach. Depth 12 is the max for Background SHAP after extensive experimenting
