+++ woodelf/explainer.py
@@ -10,7 +10,7 @@ from woodelf.core.path_to_s_vectors.base_p2s import WoodelfPathToSVectors
 from woodelf.core.trees.decision_trees_ensemble import DecisionTreesEnsemble
 from woodelf.core.trees.parse_models import load_decision_tree_ensemble_model
 from woodelf.high_depth_woodelf import woodelf_for_high_depth
-from woodelf.hybrid_woodelf import hybrid_woodelf
+from woodelf.woodelf_sparse import hybrid_woodelf
 
 AVAILABLE_MODEL_OUTPUTS = ["raw", "probability", "log_loss"]
 AVAILABLE_FEATURE_PERTURBATION = ["auto", "interventional", "tree_path_dependent"]
